+++ server/src/main/java/com/onetool/server/api/blueprint/repository/BlueprintRepository.java
@@ -8,6 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
+import java.util.List;
 
 @Repository
 public interface BlueprintRepository extends JpaRepository<Blueprint, Long> {
